requirements.txt
numpy
pandas
tqdm
matplotlib
seaborn
pillow
opencv-python
SimpleITK==2.1.1
nibabel
pydicom
seaborn
pyradiomics
scikit-learn
albumentations
imbalanced-learn==0.6.2